automate_emails/requirements.txt
numpy==2.2.3
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
tzdata==2025.1
